chore: restructure survey module folders
